tools/mq-to-bq/multiplexer_src/requirements.txt
google-cloud-pubsub
google-cloud-logging
google-cloud-asset
google-api-core
